+++ order-service/src/main/java/com/node5/orderservice/global/exception/openfeign/client/WalletClient.java
@@ -1,8 +1,8 @@
-package com.node5.orderservice.order.client;
+package com.node5.orderservice.global.exception.openfeign.client;
 
-import com.node5.orderservice.order.client.dto.WalletInfo;
-import com.node5.orderservice.order.client.dto.WalletRefundRequest;
-import com.node5.orderservice.order.client.dto.WalletWithdrawRequest;
+import com.node5.orderservice.global.exception.openfeign.client.dto.WalletInfo;
+import com.node5.orderservice.global.exception.openfeign.client.dto.WalletRefundRequest;
+import com.node5.orderservice.global.exception.openfeign.client.dto.WalletWithdrawRequest;
 import jakarta.validation.Valid;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.ResponseEntity;
