+++ order-service/src/main/java/com/node5/orderservice/global/exception/openfeign/client/dto/SettlementSourceItem.java
@@ -1,4 +1,4 @@
-package com.node5.orderservice.order.client.dto;
+package com.node5.orderservice.global.exception.openfeign.client.dto;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
