+++ order-service/src/main/java/com/node5/orderservice/global/exception/openfeign/client/dto/WalletRefundRequest.java
@@ -1,4 +1,4 @@
-package com.node5.orderservice.order.client.dto;
+package com.node5.orderservice.global.exception.openfeign.client.dto;
 
 import java.util.UUID;
