+++ order-service/src/main/java/com/node5/orderservice/order/application/OrderService.java
@@ -3,12 +3,12 @@ package com.node5.orderservice.order.application;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.node5.common.domain.PageInfoDto;
 import com.node5.common.exception.ExceptionResponseDto;
+import com.node5.orderservice.global.exception.openfeign.client.dto.*;
 import com.node5.orderservice.order.application.dto.OrderCommand;
 import com.node5.orderservice.order.application.dto.OrderCreateInfo;
 import com.node5.orderservice.order.application.dto.OrderItemCommand;
 import com.node5.orderservice.order.application.dto.OrderStatusInfo;
-import com.node5.orderservice.order.client.CatalogClient;
-import com.node5.orderservice.order.client.dto.*;
+import com.node5.orderservice.global.exception.openfeign.client.CatalogClient;
 import com.node5.orderservice.order.domain.Order;
 import com.node5.orderservice.order.domain.OrderItem;
 import com.node5.orderservice.order.domain.OrderItemRepository;
@@ -17,7 +17,7 @@ import com.node5.orderservice.order.exception.*;
 import feign.FeignException;
 import lombok.RequiredArgsConstructor;
 import com.node5.orderservice.order.application.dto.*;
-import com.node5.orderservice.order.client.WalletClient;
+import com.node5.orderservice.global.exception.openfeign.client.WalletClient;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
